// src/models/MovieModel.ts
export interface MovieModel {
  poster_path?: string;
  adult: boolean;
  overview: string;
  release_date: string;
  genereds: number[];
  id: number;
  original_title: string;
  original_language: string;
  title: string;
  backdrop_path?: string;
  popularity: number;
  vote_count: number;
  video: boolean;
  vote_average: number;
}